LayoutTests/http/tests/media/pdf-served-as-pdf-expected.txt

EXPECTED (video.error == 'null') OK
EVENT(loadstart)
EVENT(error)
failed trying to load PDF file served as PDF OK
END OF TEST
